[WIP] Keychain opt-in and increased verbosity
The use of Keychain authentication on OSX is now optionnal, and off by
default. The -k flag has therefore been added to install and
local_install which both need the elevated privileges.
Still WIP so not marked as finished.

Keychain storing and deleting are now explicit and no longer limited to
verbose.

TODO has been updated.