Properly fetch all Linux versions (#6)
* [FIX] Retrieve available versions on Linux > 5.5.0

Unity changed the listing for Linux versions above 5.5.0, and they
weren't listed properly.
This fixes this issue and also takes care of the special Linux
version naming ('0.0.0xf0')

* Parse linux versions as regular versions

The Linux installer files often came with an 'x' in front of the release
type letter. It is not meant to be permanent and u3d parses it as a
regular version by getting rid of this x whenever necessary.