[linux] Adjust to weird editor versions stored under ProjectSettings/ProjectVersion.txt (#33)
The YAML configuration file of a Unity project may often refer to weird Linux versions such as '2017.1.0xf3Linux'. To be sure that this is recognized by u3d, we parse this type of version (so in this example it would be '2017.1.0f3').

Comes with spec to make sure it works as intended.