Fix prettifier exception rule start pattern (#273)
The prettifier was considering that lines such as the following

This is a warning because of an exception (Exception: message)

should start the exception rule, while it should not. This makes it so
that the rule makes sure that there is no parenthesis on the line before
the exception to start the exception rule.